Removed the double-batching going on with the Jaeger Exporter. Added in some perf improvements.
