+++ src/OpenTelemetry/Internal/PooledList.cs
@@ -22,7 +22,6 @@ namespace OpenTelemetry.Internal
 {
     internal readonly struct PooledList<T> : IEnumerable<T>, ICollection
     {
-        private static readonly ArrayPool<T> Pool = ArrayPool<T>.Create(4096, 64);
         private static int lastAllocatedSize = 64;
 
         private readonly T[] buffer;
@@ -48,7 +47,7 @@ namespace OpenTelemetry.Internal
 
         public static PooledList<T> Create()
         {
-            return new PooledList<T>(Pool.Rent(lastAllocatedSize), 0);
+            return new PooledList<T>(ArrayPool<T>.Shared.Rent(lastAllocatedSize), 0);
         }
 
         public static void Add(ref PooledList<T> list, T item)
@@ -65,11 +64,11 @@ namespace OpenTelemetry.Internal
                 lastAllocatedSize = buffer.Length * 2;
                 var previousBuffer = buffer;
 
-                buffer = Pool.Rent(lastAllocatedSize);
+                buffer = ArrayPool<T>.Shared.Rent(lastAllocatedSize);
 
                 var span = previousBuffer.AsSpan();
                 span.CopyTo(buffer);
-                Pool.Return(previousBuffer);
+                ArrayPool<T>.Shared.Return(previousBuffer);
             }
 
             buffer[list.Count] = item;
@@ -86,7 +85,7 @@ namespace OpenTelemetry.Internal
             var buffer = this.buffer;
             if (buffer != null)
             {
-                Pool.Return(buffer);
+                ArrayPool<T>.Shared.Return(buffer);
             }
         }
